+++ packages/ldap-auth-backend/src/errors.ts
@@ -5,3 +5,6 @@ export const LDAP_CONNECT_FAIL = 'Failed to connect to LDAP server: ';
 
 export const AUTH_MISSING_CREDENTIALS =
     'AUTH_MISSING_CREDENTIALS: provide username and password or token';
+
+export const AUTH_USER_NOT_FOUND =
+    'AUTH_USER_NOT_FOUND: Credential invalid or user doesnt exists';