+++ CHANGELOG.md
@@ -3,15 +3,29 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [1.1.3](https://github.com/immobiliare/backstage-plugin-ldap-auth/compare/v1.1.2...v1.1.3) (2022-11-11)
+# [2.0.0-alpha.2](https://github.com/immobiliare/backstage-plugin-ldap-auth/compare/v2.0.0-alpha.1...v2.0.0-alpha.2) (2022-11-21)
+
+### Bug Fixes
+
+-   **ldap.ts:** default auth function signature ([3f2dd02](https://github.com/immobiliare/backstage-plugin-ldap-auth/commit/3f2dd0270dd1fd137d28a81101943d0014bf32bc))
+
+# [2.0.0-alpha.1](https://github.com/immobiliare/backstage-plugin-ldap-auth/compare/v2.0.0-alpha.0...v2.0.0-alpha.1) (2022-11-21)
+
+**Note:** Version bump only for package root
+
+# [2.0.0-alpha.0](https://github.com/immobiliare/backstage-plugin-ldap-auth/compare/v1.1.2-alpha.0...v2.0.0-alpha.0) (2022-11-17)
 
 ### Bug Fixes
 
 -   allow no children to the LdapAuthFrontendPage ([7624203](https://github.com/immobiliare/backstage-plugin-ldap-auth/commit/7624203260d6201815a1f8c79a4be98ba880cb89))
 
-## [1.1.2](https://github.com/immobiliare/backstage-plugin-ldap-auth/compare/v1.1.2-alpha.0...v1.1.2) (2022-10-28)
+### Features
 
-**Note:** Version bump only for package root
+-   reworked options and injectables ([b6b1445](https://github.com/immobiliare/backstage-plugin-ldap-auth/commit/b6b144539827301a7c95d6547d0217a1fe3cc103))
+
+### BREAKING CHANGES
+
+-   Structure and options
 
 ## [1.1.2-alpha.0](https://github.com/immobiliare/backstage-plugin-ldap-auth/compare/v1.1.1...v1.1.2-alpha.0) (2022-10-28)